requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pandas>=2.2.0
openpyxl>=3.1.2
pydantic>=2.5.0
python-multipart>=0.0.6
sqlmodel>=0.0.14
xhtml2pdf>=0.2.13
jinja2>=3.1.2
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
